sched/ghost: split out bpf helper ID definitions
enum bpf_func_id is created by a macro that is used for an array lookup
from helper ID to string (func_id_str in disasm.c).

We don't want that array to be too big.  However, we also don't want our
helper ID numbers to clash with upstream's function IDs.  That will
happen eventually.

Move the ghost bpf helper IDs out of bpf_func_id so that we can use a
larger number without affecting the array.

The array mapping is optional - it's used in func_id-to-name lookups,
and those lookups are protected in case the func_id > __BPF_FUNC_MAX_ID.

Effort: sched/ghost

+++ kernel/sched/ghost_core.c
@@ -1154,7 +1154,8 @@ ghost_sched_func_proto(enum bpf_func_id func_id, const struct bpf_prog *prog)
 {
 	int eat = bpf_prog_eat_type(prog);
 
-	switch (func_id) {
+	/* ghost func_ids are not in enum bpf_func_id */
+	switch ((int)func_id) {
 	case BPF_FUNC_ghost_wake_agent:
 		return &bpf_ghost_wake_agent_proto;
 	case BPF_FUNC_ghost_run_gtid:
@@ -1191,7 +1192,8 @@ const struct bpf_prog_ops ghost_sched_prog_ops = {};
 static const struct bpf_func_proto *
 ghost_msg_func_proto(enum bpf_func_id func_id, const struct bpf_prog *prog)
 {
-	switch (func_id) {
+	/* ghost func_ids are not in enum bpf_func_id */
+	switch ((int)func_id) {
 	case BPF_FUNC_ghost_wake_agent:
 		return &bpf_ghost_wake_agent_proto;
 	case BPF_FUNC_ghost_resched_cpu:
